requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
xarray>=2023.10.0
netcdf4>=1.6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# AI/ML
openai>=1.3.0
langchain>=0.0.300
langchain-openai>=0.0.1

# Visualization
plotly>=5.17.0
folium>=0.15.0
pydeck>=0.8.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0

# Ocean science
gsw>=3.6.0
cftime>=1.6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
pydantic-settings>=2.0.0
